move precache into hw_cache.c
