+++ src/hardware/hw_glob.h
@@ -49,13 +49,16 @@ typedef struct gl_vissprite_s
 extern extrasubsector_t *extrasubsectors;
 extern size_t addsubsector;
 
+void HWR_FreeExtraSubsectors(void);
+
 // --------
 // hw_cache.c
 // --------
 void HWR_InitTextureCache(void);
 void HWR_FreeTextureCache(void);
 void HWR_FreeMipmapCache(void);
-void HWR_FreeExtraSubsectors(void);
+
+void HWR_PrecacheLevel(void);
 
 void HWR_GetFlat(lumpnum_t flatlumpnum, boolean noencoremap);
 // ^ some flats must NOT be remapped to encore, since we remap them as we cache them for ease, adding a toggle here seems wise.
